Nest iOS microphone CLI actions
